Implemented Adding and Deleting Recipes

+++ Grocery/ListTableViewController.swift
@@ -18,6 +18,7 @@ class ListTableViewController: UITableViewController {
     var editDate: NSString = ""
     var editId: Int32?
     var recipeList: [String] = []
+    var recipeTitle: String?
     var label: String?
     let queryFoodStatementString = "SELECT * FROM Food ORDER BY food ASC;"
     let queryRecipeStatementString = "SELECT * FROM Recipes ORDER BY name ASC;"
@@ -88,6 +89,12 @@ class ListTableViewController: UITableViewController {
             }
             
         }
+        if segue.destination is DisplayRecipeTableViewController{
+            let vc = segue.destination as? DisplayRecipeTableViewController
+            vc?.db = db
+            vc?.label = label
+            vc?.recipeTitle = recipeTitle
+        }
     }
  
 
@@ -123,9 +130,12 @@ class ListTableViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == UITableViewCell.EditingStyle.delete{
             
-            let deleteStatementString = "DELETE FROM Food WHERE food = '\(foodList[indexPath.row])' AND date = '\(dateList[indexPath.row])';"
+            if label == "Food"{
+                let deleteStatmentString = "DELETE FROM Food WHERE food = '\(foodList[indexPath.row])' AND date = '\(dateList[indexPath.row])';"
+            
+          
             var deleteStatement: OpaquePointer? = nil
-            if sqlite3_prepare_v2(db, deleteStatementString, -1, &deleteStatement, nil) == SQLITE_OK {
+            if sqlite3_prepare_v2(db, deleteStatmentString, -1, &deleteStatement, nil) == SQLITE_OK {
                 if sqlite3_step(deleteStatement) == SQLITE_DONE {
                     print("Successfully deleted row.")
                 } else {
@@ -139,7 +149,58 @@ class ListTableViewController: UITableViewController {
             dateList.remove(at: indexPath.row)
             tableView.reloadData()
 
-
+            }
+            else{
+                var id: Int32 = 0
+                id = GetId(recipeName: recipeList[indexPath.row])
+                print(id)
+               let deleteStatmentString = "DELETE FROM Recipes WHERE recipeId = '\(id)';"
+                let deleteIngredientStatementString = "DELETE FROM Ingredients WHERE recipeId = '\(id)';"
+                let deleteStepsStatementString = "DELETE FROM Steps WHERE recipeId = '\(id)';"
+                
+                //Delete Steps
+                var deleteStepStatement: OpaquePointer? = nil
+                if sqlite3_prepare_v2(db, deleteStepsStatementString, -1, &deleteStepStatement, nil) == SQLITE_OK {
+                    if sqlite3_step(deleteStepStatement) == SQLITE_DONE {
+                        print("Successfully deleted step row.")
+                    } else {
+                        print("Could not delete step row.")
+                    }
+                } else {
+                    print("DELETE statement could not be prepared")
+                }
+                sqlite3_finalize(deleteStepStatement)
+                
+                //Delete Ingredients
+                var deleteIngredientStatement: OpaquePointer? = nil
+                
+                if sqlite3_prepare_v2(db, deleteIngredientStatementString, -1, &deleteIngredientStatement, nil) == SQLITE_OK {
+                    if sqlite3_step(deleteIngredientStatement) == SQLITE_DONE {
+                        print("Successfully deleted Ingredient row.")
+                    } else {
+                        print("Could not delete ingredient row.")
+                    }
+                } else {
+                    print("DELETE statement could not be prepared")
+                }
+                sqlite3_finalize(deleteIngredientStatement)
+               
+                //Delete Recipe
+                var deleteStatement: OpaquePointer? = nil
+                if sqlite3_prepare_v2(db, deleteStatmentString, -1, &deleteStatement, nil) == SQLITE_OK {
+                    if sqlite3_step(deleteStatement) == SQLITE_DONE {
+                        print("Successfully deleted row.")
+                    } else {
+                        print("Could not delete row.")
+                    }
+                } else {
+                    print("DELETE statement could not be prepared")
+                }
+                sqlite3_finalize(deleteStatement)
+                recipeList.remove(at: indexPath.row)
+                tableView.reloadData()
+                
+            }
         }
         
     }
@@ -195,6 +256,28 @@ class ListTableViewController: UITableViewController {
         editDate = dateList[indexPath.row] as NSString
         performSegue(withIdentifier: "editSegue", sender: self)
         }
+        else if label == "Recipes"{
+            recipeTitle = recipeList[indexPath.row]
+            performSegue(withIdentifier: "displayRecipeSegue", sender: self)
+        }
+    }
+    
+    func GetId(recipeName: String) -> Int32 {
+        var newId: Int32 = 0
+        let queryIdStatementString = "SELECT recipeId FROM Recipes WHERE name = '\(recipeName)';"
+        var queryIdStatement: OpaquePointer? = nil
+        if sqlite3_prepare_v2(db, queryIdStatementString, -1, &queryIdStatement, nil) != SQLITE_OK{
+            print("Error binding get Id query")
+        }
+        
+        while (sqlite3_step(queryIdStatement) == SQLITE_ROW){
+            newId = sqlite3_column_int(queryIdStatement, 0)
+        }
+        
+        return newId
+       
+        
+        
     }
